+++ src/game/CMakeLists.txt
@@ -246,6 +246,7 @@ set (game_SRCS
     ObjectGuid.h
     ObjectMgr.h
     ObjectPosSelector.h
+    PacketProcessing.h
     PlayerDump.h
     QuestDef.h
     ReputationMgr.h
@@ -390,6 +391,7 @@ set (game_SRCS
     Objects/GameObject.h
     Objects/GameObjectDefines.h
     Objects/Item.h
+    Objects/ItemDefines.h
     Objects/ItemPrototype.h
     Objects/MirrorTimer.h
     Objects/MovementInfo.h
