+++ src/game/HardcodedEvents.cpp
@@ -5,13 +5,14 @@
 #include "HardcodedEvents.h"
 #include "World.h"
 #include "MapManager.h"
+#include "WaypointManager.h"
+#include "Chat.h"
+#include "GridSearchers.h"
 #include "world/scourge_invasion.h"
 #include "world/world_event_wareffort.h"
-#include "GridSearchers.h"
 #include <chrono>
 #include <random>
 #include <limits>
-#include "WaypointManager.h"
 
 /*
  * Elemental Invasion
