+++ src/game/Maps/MapManager.h
@@ -25,10 +25,10 @@
 #include "Common.h"
 #include "Platform/Define.h"
 #include "Policies/Singleton.h"
-#include "Map.h"
 #include "GridStates.h"
 #include <condition_variable>
 
+class Map;
 class BattleGround;
 
 enum
@@ -145,21 +145,6 @@ class MapManager : public MaNGOS::Singleton<MapManager, MaNGOS::ClassLevelLockab
         static bool IsValidMapCoord(WorldLocation const& loc)
         {
             return IsValidMapCoord(loc.mapId,loc.x,loc.y,loc.z,loc.o);
-        }
-
-        // modulos a radian orientation to the range of 0..2PI
-        static float NormalizeOrientation(float o)
-        {
-            // fmod only supports positive numbers. Thus we have
-            // to emulate negative numbers
-            if (o < 0)
-            {
-                float mod = o *-1;
-                mod = fmod(mod, 2.0f*M_PI_F);
-                mod = -mod+2.0f*M_PI_F;
-                return mod;
-            }
-            return fmod(o, 2.0f*M_PI_F);
         }
 
         void RemoveAllObjectsInRemoveList();
