+++ src/game/Movement/PointMovementGenerator.cpp
@@ -19,9 +19,12 @@
 #include "PointMovementGenerator.h"
 #include "Creature.h"
 #include "CreatureAI.h"
+#include "GameObject.h"
 #include "GameObjectAI.h"
+#include "Player.h"
 #include "PlayerAI.h"
 #include "TemporarySummon.h"
+#include "Map.h"
 #include "World.h"
 #include "MoveSplineInit.h"
 #include "MoveSpline.h"
