+++ src/game/Objects/Unit.cpp
@@ -49,11 +49,11 @@
 #include "MovementGenerator.h"
 #include "Transport.h"
 #include "CreatureGroups.h"
+#include "ScriptMgr.h"
 #include "ZoneScript.h"
 #include "MoveSplineInit.h"
 #include "MoveSpline.h"
 #include "packet_builder.h"
-#include "Chat.h"
 #include "Anticheat.h"
 #include "InstanceStatistics.h"
 #include "MovementPacketSender.h"
