+++ src/game/OutdoorPvP/OutdoorPvPSI.cpp
@@ -23,6 +23,7 @@
 #include "Language.h"
 #include "World.h"
 #include "GameEventMgr.h"
+#include "ScriptMgr.h"
 
 #if SUPPORTED_CLIENT_BUILD > CLIENT_BUILD_1_11_2
