+++ src/scripts/PrecompiledHeaders/scriptPCH.h
@@ -26,7 +26,6 @@
 #include "SpellAuras.h"
 #include "TemporarySummon.h"
 #include "NullCreatureAI.h"
-#include "Chat.h"
 #include "BattleGroundAV.h"
 #include "World.h"
 #include "Weather.h"
