+++ src/scripts/custom/custom_creatures.cpp
@@ -898,7 +898,6 @@ void LearnSkillRecipesHelper(Player *player, uint32 skill_id)
 
 bool LearnAllRecipesInProfession(Player *pPlayer, SkillType skill)
 {
-    ChatHandler handler(pPlayer->GetSession());
     char* skill_name;
 
     SkillLineEntry const *SkillInfo = sSkillLineStore.LookupEntry(skill);
